refine

+++ include/tsv.h
@@ -56,7 +56,7 @@ namespace utils
 
 	inline std::vector<std::string> split_by_cnt(const std::string &s, char delim, int fixcnt) {
 		auto elems = split_by_cuts(s, delim);
-		while (elems.size() < fixcnt)
+		while ((int)elems.size() < fixcnt)
 			elems.push_back("");
 		return elems;
 	}
